requirements.txt
PyYAML==6.0

fastapi==0.103.2
uvicorn==0.23.2

requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.2
pydantic==2.4.2

#PDF extract tool
pdfminer.six==20231228

retry==0.9.2

typer==0.9.0
schedule==1.2.1

langchain==0.0.305
# langchainplus-sdk==0.0.20
langsmith==0.0.41
qdrant-client==1.5.4
supabase==1.0.2
pinecone-client==2.2.2
sentence_transformers==2.2.2
openai==1.3.8
tavily-python==0.2.9

sendgrid==6.10.0

# Clean code tools
black==23.9.1
isort==5.12.0

# Evaluation
# ragas==0.1.0rc1
